hide remove section link for users; fix description refresh issue

+++ app/assets/javascripts/discussions.js
@@ -1,4 +1,7 @@
 function rejectDiscussion(id) {
+    var response = confirm('Are you sure?');
+
+    if (!response) { return; }
     $.ajax({
         url: '/discussions/'+id,
         type: 'DELETE',
